refactor: change file structure of member domain
