+++ src/main/java/com/curateme/claco/member/domain/entity/Member.java
@@ -1,4 +1,4 @@
-package com.curateme.claco.member.entity;
+package com.curateme.claco.member.domain.entity;
 
 import com.curateme.claco.global.entity.BaseEntity;
