+++ src/main/java/com/curateme/claco/member/domain/entity/Role.java
